added decryption, added check to prevent duplication in random number, added set map to set the key into charmap.

+++ encryption/CharMap.go
@@ -1,6 +1,6 @@
 package encryption
 
-var CharMap = map[string]int{
+var CharMap = map[string]int64{
 	"a": 0, "b": 0, "c": 0, "d": 0, "e": 0, "f": 0,
 	"g": 0, "h": 0, "i": 0, "j": 0, "k": 0, "l": 0,
 	"m": 0, "n": 0, "o": 0, "p": 0, "q": 0, "r": 0,
